fix: some route & add connection pooling

+++ flask/src/service/auth_service.py
@@ -9,11 +9,16 @@ from validation.validation import validate
 
 def login(req):
     cursor = get_cursor()
+    if cursor is None:
+        raise ResponseError(500, "Error obtaining database connection")
+    
     try:
         data = validate(LoginUserValidation, req)
 
         email = data.email
         password = data.password
+
+        print(email)
 
         cursor.execute(
             "SELECT id, email, nama, role_id, password FROM users WHERE email = %s", 
@@ -21,11 +26,13 @@ def login(req):
         )
         user = cursor.fetchone()
 
+        print(user)
+
         if not user:
-            raise ResponseError(400, "Username or password wrong")
+            raise ResponseError(400, "Username or password wrong hehe")
 
         if not bcrypt.checkpw(password.encode('utf-8'), user['password'].encode('utf-8')):
-            raise ResponseError(400, "Username or password wrong")
+            raise ResponseError(400, "Username or password wrong haha")
 
         exp_time = (datetime.now() + timedelta(hours=2)).timestamp()
 
@@ -43,12 +50,12 @@ def login(req):
 
         cursor.execute(
             """
-            INSERT INTO sessions (token, email, expiry, created_at, updated_at)
+            INSERT INTO sessions (token, user_id, expiry, created_at, updated_at)
             VALUES (%s, %s, %s, %s, %s)
             """, 
             (
                 token,
-                user['email'],
+                user['id'],
                 datetime.utcnow() + timedelta(hours=2),
                 datetime.utcnow(),
                 datetime.utcnow()
@@ -61,9 +68,16 @@ def login(req):
         raise e
     except Exception as e:
         raise ResponseError(500, str(e))
+    finally:
+        if cursor:
+            cursor.close()
+    
     
 def logout(token):
     cursor = get_cursor()
+    if cursor is None:
+        raise ResponseError(500, "Error obtaining database connection")
+    
     try:
         cursor.execute(
             "DELETE FROM sessions WHERE token = %s", 
@@ -76,3 +90,6 @@ def logout(token):
         raise e 
     except Exception as e:
         raise ResponseError(500, str(e))
+    finally:
+        if cursor:
+            cursor.close()
